+++ tests/conftest.py
@@ -1,11 +1,11 @@
-import Tracker.db as db
+from Tracker import db
 import pytest
 import contextlib
 
 
 @pytest.fixture(scope="function", autouse=True)
 def do_something(request):
-    meta = db.sqlBase.metadata
+    meta = db.metadata
 
     with contextlib.closing(db.engine.connect()) as con:
         trans = con.begin()
